Create README.md
I studied Bayesian methods 35 years ago as an undergraduate in Durham with Julian Besag and then at UCL with Philip Dawid . I then took shelter from the second AI winter spending 20 years as a reinsurance underwriter and actuary in the London Market. (A chance conversation when I worked as a barman after uni and a very liquid lunch/interview led me to this career path)

Some of their teaching must have sunk in because, looking back, I had developed a successful method for risk selection and built a profitable portfolio able to absorb shock losses thinking as a Bayesian, not a frequentist.

So now I am looking to update my Bayesian toolkit and, as a domain expert in Reinsurance myself, create bridges between a very traditional and sceptical profession and the AI industry

Think deeply about a risk, stress less.